ContactPhotoManager: Consider BITMAP_UNAVAILABLE as a cache miss
When a default image is generated, it is cached as an instance of
BITMAP_UNAVAILABLE. The code that renders default images interprets
null (not present) and null content (invalid byte array) as misses
and generates default images. However, it misses the case of
previously generated default images.

An easy way to repro this is in Sim Contacts:
1) Go to contacts
2) Chose Import from Sim
3) Observe that images are shown correctly
4) Go back, chose import from sim again
5) Observe that images are missing
6) (As a side-effect pressing one will crash the app).

This patch considers a cache BITMAP_UNAVAILABLE as a cache miss
when generating default images, fixing the problem above.

Change-Id: I61050e68b27c605392d0649316ba12a4660a7289
Ticket: CYNGNOS-1829
Ticket: CRACKLING-910
Ticket: KIPPER-590